Repair and Offline Mode: mouse box-dragging function application in SP
See #17339. Allow players to drag a box to repair and offline/online buildings and units, as well as originally clicking one by one.

Make that little wrench and tiny spark more convenient and helpful to players.

Repair Selection Priority is modified due to that.
-- Default: buildings > combat vehicles and aircrafts > non-combat vehicles and aircrafts > allied buildings
-- Ctrl: all
-- Alt: combat vehicles and aircrafts > non-combat vehicles and aircrafts > buildings > allied buildings

New Modifiers mode For Powerdown Mode:
- Default: generally turn on/off with smart selection.
- Ctrl: Only turn off.
- Alt: Only turn on.